+++ src/crypto/aes/mod.rs
@@ -1,4 +1,3 @@
 pub mod aes;
 pub mod encryption;
-pub mod errors;
 pub mod key;
